Hide delete button after deleting a node.
